dis88, from dev86, I think, with hacks for disassembling VENIX files instead of PC/IX files. Though they are both derived from V7, they are different in some details that are TBD.
The original ran on pc/ix, but this one won't run on Venix. I've
included copies of the venix include files hacked to have machine
independent types.

Also included a bunch of heuristics that attempt to make the
disassembly better for the Venix C compiler.

Note: 8087 op-codes are not yet disassembled, and this gets confused
about data symbols as well sometimes (eg _lbolt+3255) for reasons
unknown.