Implement Venix compiler work arounds
So, The Venix toolchain does weird things sometimes. It always starts
'text' from a new file at an even address, with zero padding. This
zero padding confuses the disassembler because 0x00 is a valid
instruction that then interprets the following push bp as part of that
instruction.  Instead, implement a heuristic that if we're at an odd
address, and the last instruction was a ret and the current byte is a
0, emit .byte 0 instead. This likely should be a quirk if we ever want
to target PC/IX or any of the other historic binary-only unix
ports. For now, though, it's not.