wutil32.flp: wutil 3.2 for DEC Rainbow
wutil 3.2 is the last version of the Winchester configuration utility
for the DEC Rainbow. This MS-DOS bootable floppy contains DOS and
wutil32. You use it to low level format blank disks, partition
disks and set partitions bootable. You'll need this to setup
any hard disk (real or virtual) that you want to install
Venix onto. It will be part of the instructions.