Fix 0xe9 jumps
ip + fetchWord() is an order of operations problem. fetchWord will
increment ip by two, so it's not well defined what should
happen. Rewrite in a way that is well defined. It appres to be the
only such place we do this.

This fixes all manner of crazy things (including cal).