Make high stack binaries work.
There was a check for bad addresses in write. It should be in a lot of
other places, but it's not right now. It assumed that brk was the
highest valid address. It's not, the valid addresses are between 0 and
brk and sp and ffff when we have high-stack binaries. Update the test
to reflect that.

Update some socks. Save the stack size in the loader along with the
magic in case we need that elsewhere.