Implement fork.
But this shows three issues:

(1) errno needs to be mapped becaue EAGAIN differs between FreeBSD and
    V7. So we need to translate host errno to venix. Likely this
    can just be done in sys_error()
(2) fork creates a new address space, so the whole pid tracking thing
    is lame. But it may be the right lies to not matter.
(3) We have no SIGCHLD handler, and we need that for wait(), et al.

But this gets us a little closer to cc working: now wait and exece
are complained about. Tiny progress.