First attempt to reconstruct crt0.s
This assembles to nearly the same as the original crt0.o. However, there's some
differences in the relocation elements, for some reason. dis88 gives the same
output of the two different .o's, but hexdump shows differences in the reloc
section for reasons unknown.

Note: this is with the VENIX /bin/as running under my hacked up 86sim+venix
system calls thing also in this repo. For the two or three trivial test cases I
have, running like this produces identical output to the as running on the
Rainbow, so any differences at all need to be investigated when I have some
time.