Report the unused padding for the relocation data.
When we build crt0.o from crt0.s with the assembler in emulation, we get
different numbers than what's shipped with Venix. It's not clear why, but this
is what the differences are, but this makes them visible to see...

When I do:
% V=$HOME/venix/extracted
% ./86sim $V/bin/as ../../src/libc/csu/crt0.s -o foo.o
% ./86sim $V/bin/ld foo.o hello.o $V/lib/libc.a -o hello-foo
% ./86sim hello-foo
Hello world
% ./86sim $V/bin/ld $V/lib/crt0.o hello.o $V/lib/libc.a -o hello-org
% ./86sim hello-org
Hello world
%

so I'm not sure the differences really matter here. We'll see on larger
programs. The programs are one byte different in size, though:
-rwxr-xr-x  1 imp  imp  9496 Apr 19 11:13 hello-foo
-rwxr-xr-x  1 imp  imp  9497 Apr 19 11:18 hello-org
so maybe there is a difference? If I copy foo.o -> foo1.o and $V/lib/crt0.o
to foo2.o, then the size is the same, but they differ:
-00001fd0  05 01 1b 04 66 12 00 00  41 05 66 6f 6f 31 2e 6f  |....f...A.foo1.o|
+00001fd0  05 01 1b 04 66 12 00 00  41 05 66 6f 6f 32 2e 6f  |....f...A.foo2.o|
so I think that's well within the 'It's fine' neighborhood and the differences
don't matter and are likely an artifiact of how $V/lib/crt0.o was built that's
different in emulation and/or maybe a different assembler version...

ALSO: Need to start writing some kind of regression tests for 86sim now that
it's working decently well.