Add the profile version of the startup code.
From the v7 sources as well as cribbed data from the disassembly.

As with the crt0.s file, this produces almost identical results to the mcrt0.o
supplied with Venix/86, modulo the 'unused' bits in the text reloc entries.

This was a bit more tricky because the code is more complicated and a number of
unfamiliar constructs were needed. The v7 code has little to no comments, so
I've continued that there, though I should likely add them.

A number of stylistic things were inspired from whatever uboot source code does.