Implemnet times(2) system call
This returns the amount of consumed time and the system uptime.  We
assume Hz on the host is 1000 and 60 on the emulated system.  The
Venix manual says the units are 1/Hz seconds, where Hz is 60, but I
believe that this would be 50 in Europe since the Rainbow's internal
periodic interrupt ran at 50Hz there.