Add docs for exec
Add docs for exec. Looking at the /bin/sh sources, we're putting some
kind of garbage into the env and it's barfing when it starts up. Since
we don't implement the env, this makes sense. We're likely doing
something wrong in exec and that was rejiggered recently so /bin/sh
failing kinda makes some sense.