wait and exec enhancements
Only return 2 bytes of status for wait.
Make sure that we return the right error when exec fails if the args are OK.

This lets cc orchestrate the build except for the optimizer. It's unclear
why that's failing. Run on its own, it works. And substituting /bin/cat
also fails.

/bin/csh returns signal 38 error status for commands that have otherwise
run to completion. It's unclear why this would be the case.