wait: Fix return value
The value for wait gets returned in %dx. The code had assumed that
%ds:%dx contained a pointer to where to return the status. In reality,
the status is returned in %dx itself and the glue code writes it into
the passed in addres. Set EDX to be the status value instead of copying
out the value to what really isn't a pointer....

This fixes cc -O and many other weird things I'd seen (commands in
/bin/csh returning with a weird signal, etc).